fix: use enhanced PATH when spawning agent processes
The agent check used an enhanced PATH (including ~/.opencode/bin) to
find commands, but AgentProcess.spawn() used the default PATH. This
caused ENOENT errors when creating sessions for agents installed in
custom locations.

Extract getEnhancedPath() to shared utils/path.ts and use it in both
agent-check.ts and AgentProcess.ts for consistent command resolution.
